backend/py/requirements.txt
sqlalchemy==1.4.46
fastapi==0.89.1
uvicorn[standard]==0.20.0
alembic==1.9.2
tenacity==8.1.0
pydantic==1.10.4
pydantic[email]
psycopg2
passlib==1.7.4
scrypt==0.8.20
python-jose==3.3.0
httpx==0.23.3
requests==2.28.2
lxml==4.9.2
pytest==7.2.1